Move both issuedVc and validVc to json.
