+++ index.js
@@ -8,8 +8,13 @@ import {
   shouldBeBs58, verificationFail
 } from './helpers.js';
 import chai from 'chai';
+import {createRequire} from 'node:module';
 import {generateTestData} from './vc-generator/index.js';
-import {validVc} from './validVc.js';
+
+//FIXME remove this once node has non-experimental support
+//for es6 import and json
+const require = createRequire(import.meta.url);
+export const validVc = require('./validVc.json');
 
 const should = chai.should();
