+++ vc-generator/defaultKey.json
@@ -1,0 +1,8 @@
+{
+  "id": "did:key:z6MkwXG2WjeQnNxSoynSGYU8V9j3QzP3JSqhdmkHc6SaVWoT#z6MkwXG2WjeQnNxSoynSGYU8V9j3QzP3JSqhdmkHc6SaVWoT",
+  "@context": "https://w3id.org/security/multikey/v1",
+  "type": "Multikey",
+  "controller": "did:key:z6MkwXG2WjeQnNxSoynSGYU8V9j3QzP3JSqhdmkHc6SaVWoT",
+  "publicKeyMultibase": "z6MkwXG2WjeQnNxSoynSGYU8V9j3QzP3JSqhdmkHc6SaVWoT",
+  "secretKeyMultibase": "zrv3rbPamVDGvrm7LkYPLWYJ35P9audujKKsWn3x29EUiGwwhdZQd1iHhrsmZidtVALBQmhX3j9E5Fvx6Kr29DPt6LH"
+}
